solve: temp 변수의 산술 오버플로 발견 및 수정

+++ jlis/solution.cpp
@@ -38,7 +38,7 @@ int calcLis(int ni, int mi)
         return ret;
 
     ret = 2;
-    int cur = std::max(nArr[ni], mArr[mi]);
+    long long cur = std::max(nArr[ni], mArr[mi]);
 
     FOR(nxt, ni + 1, N + 1)
     if (nArr[nxt] > cur)
